pyproject.toml
[project]
name = "stegcipher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
]
